requirements.txt
openenv-core>=0.2.3
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
requests>=2.31.0
openai>=1.20.0
pyyaml>=6.0
